refactor: 페키지 구조 개선
