+++ src/main/java/blackjack/domain/game/GameReferee.java
@@ -1,0 +1,8 @@
+package blackjack.domain.game;
+
+import blackjack.domain.participants.Dealer;
+import blackjack.domain.participants.Player;
+
+public interface GameReferee {
+    GameResult judge(Dealer dealer, Player player);
+}
