+++ src/main/java/blackjack/domain/game/GameResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain.result;
+package blackjack.domain.game;
 
 public enum GameResult {
     PLAYER_WIN, DEALER_WIN, PUSH
