--- src/main/java/blackjack/domain/result/GameReferee.java
@@ -1,8 +1,0 @@
-package blackjack.domain.result;
-
-import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Player;
-
-public interface GameReferee {
-    GameResult judge(Dealer dealer, Player player);
-}
