+++ src/main/java/blackjack/dto/DealerHitDto.java
@@ -1,6 +1,6 @@
 package blackjack.dto;
 
-import blackjack.domain.participant.Dealer;
+import blackjack.domain.participants.Dealer;
 
 public record DealerHitDto(String dealerName, int hitCount) {
     public static DealerHitDto of(Dealer dealer, int dealerHitCount) {
