+++ src/main/java/blackjack/dto/ParticipantScoreDtos.java
@@ -1,8 +1,8 @@
 package blackjack.dto;
 
-import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Participant;
-import blackjack.domain.participant.Player;
+import blackjack.domain.participants.Dealer;
+import blackjack.domain.participants.Participant;
+import blackjack.domain.participants.Player;
 import java.util.List;
 import java.util.stream.Stream;
